requirements.txt
asgiref==3.5.2
azure-common==1.1.28
azure-core==1.25.0
azure-identity==1.10.0
azure-keyvault-secrets==4.4.0
azure-storage-blob==12.13.1
backports.zoneinfo==0.2.1
Django==4.1
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
python-dotenv==0.20.0
python-environ==0.4.54
sqlparse==0.4.2
urllib3==1.26.11
